requirements.txt
flask==2.3.3
openai==1.76.0
python-dotenv==1.0.0
Pillow==10.0.0